Organization- Seprerate reused functions.
